+++ Source/IO/WinCopies.IO.Shared/BitmapSources/BitmapSourcesLinker.cs
@@ -23,8 +23,19 @@ using System.Windows.Media.Imaging;
 
 namespace WinCopies.IO
 {
+    public enum BitmapSourcesLinkerLoadStep
+    {
+        Default = 0,
+
+        Intermediate = 1,
+
+        Completed = 2
+    }
+
     public interface IBitmapSourcesLinker : IBitmapSources
     {
+        IBitmapSourcesLinkerLoadSteps LoadSteps { get; }
+
         void LoadSmall();
 
         void LoadMedium();
@@ -46,31 +57,75 @@ namespace WinCopies.IO
         void OnBitmapSourcesLoaded();
     }
 
+    public interface IBitmapSourcesLinkerLoadSteps
+    {
+        BitmapSourcesLinkerLoadStep Small { get; }
+
+        BitmapSourcesLinkerLoadStep Medium { get; }
+
+        BitmapSourcesLinkerLoadStep Large { get; }
+
+        BitmapSourcesLinkerLoadStep ExtraLarge { get; }
+    }
+
+    public struct BitmapSourcesLinkerLoadSteps
+    {
+        public BitmapSourcesLinkerLoadStep Small;
+
+        public BitmapSourcesLinkerLoadStep Medium;
+
+        public BitmapSourcesLinkerLoadStep Large;
+
+        public BitmapSourcesLinkerLoadStep ExtraLarge;
+    }
+
+    public struct _BitmapSourcesLinkerLoadSteps : IBitmapSourcesLinkerLoadSteps
+    {
+        internal BitmapSourcesLinkerLoadSteps _loadSteps;
+
+        public BitmapSourcesLinkerLoadStep Small => _loadSteps.Small;
+
+        public BitmapSourcesLinkerLoadStep Medium => _loadSteps.Medium;
+
+        public BitmapSourcesLinkerLoadStep Large => _loadSteps.Large;
+
+        public BitmapSourcesLinkerLoadStep ExtraLarge => _loadSteps.ExtraLarge;
+    }
+
+    public struct BitmapSourcesLinkerStruct
+    {
+        public BitmapSource Small;
+        public BitmapSource Medium;
+        public BitmapSource Large;
+        public BitmapSource ExtraLarge;
+    }
+
     public class BitmapSourcesLinker : BitmapSources<IBitmapSourceProvider>, IBitmapSourcesLinker, INotifyPropertyChanged
     {
         private const BindingFlags _flags = BindingFlags.Public | BindingFlags.Instance;
+        private _BitmapSourcesLinkerLoadSteps _loadSteps;
+        private BitmapSourcesLinkerStruct _bitmapSources;
 
-        private BitmapSource _small;
-        private BitmapSource _medium;
-        private BitmapSource _large;
-        private BitmapSource _extraLarge;
+        protected override BitmapSource SmallOverride => _bitmapSources.Small;
 
-        protected override BitmapSource SmallOverride => _small;
+        protected override BitmapSource MediumOverride => _bitmapSources.Medium;
 
-        protected override BitmapSource MediumOverride => _medium;
+        protected override BitmapSource LargeOverride => _bitmapSources.Large;
 
-        protected override BitmapSource LargeOverride => _large;
+        protected override BitmapSource ExtraLargeOverride => _bitmapSources.ExtraLarge;
 
-        protected override BitmapSource ExtraLargeOverride => _extraLarge;
+        public _BitmapSourcesLinkerLoadSteps LoadSteps => _loadSteps;
+
+        IBitmapSourcesLinkerLoadSteps IBitmapSourcesLinker.LoadSteps => LoadSteps;
 
         public event PropertyChangedEventHandler PropertyChanged;
 
         public BitmapSourcesLinker(in IBitmapSourceProvider bitmapSourceProvider) : base(bitmapSourceProvider)
         {
-            UpdateBitmapSource(ref _small, bitmapSourceProvider.Default.Small);
-            UpdateBitmapSource(ref _medium, bitmapSourceProvider.Default.Medium);
-            UpdateBitmapSource(ref _large, bitmapSourceProvider.Default.Large);
-            UpdateBitmapSource(ref _extraLarge, bitmapSourceProvider.Default.ExtraLarge);
+            UpdateBitmapSource(ref _bitmapSources.Small, bitmapSourceProvider.Default.Small);
+            UpdateBitmapSource(ref _bitmapSources.Medium, bitmapSourceProvider.Default.Medium);
+            UpdateBitmapSource(ref _bitmapSources.Large, bitmapSourceProvider.Default.Large);
+            UpdateBitmapSource(ref _bitmapSources.ExtraLarge, bitmapSourceProvider.Default.ExtraLarge);
         }
 
         protected void UpdateBitmapSource(ref BitmapSource value, in BitmapSource newValue) => (value = newValue).Freeze();
@@ -89,15 +144,53 @@ namespace WinCopies.IO
 
         public virtual void OnBitmapSourcesLoaded() => InvokeMethods(name => name != nameof(OnBitmapSourcesLoaded) && name.StartsWith("On") && name.EndsWith("Loaded"));
 
-        public void LoadSmall() => UpdateBitmapSource(ref _small, InnerObject.Sources.Small);
+        protected void Run(in Action action, in BitmapSourcesLinkerLoadStep? loadStep)
+        {
+            if (loadStep.HasValue && loadStep.Value == BitmapSourcesLinkerLoadStep.Completed)
 
-        public void LoadMedium() => UpdateBitmapSource(ref _medium, InnerObject.Sources.Medium);
+                return;
 
-        public void LoadLarge() => UpdateBitmapSource(ref _large, InnerObject.Sources.Large);
+            action();
+        }
 
-        public void LoadExtraLarge() => UpdateBitmapSource(ref _extraLarge, InnerObject.Sources.ExtraLarge);
+        protected void Run2(ref BitmapSource value, Func<IBitmapSources, BitmapSource> selector, ref BitmapSourcesLinkerLoadStep loadStep) => WinCopies.UtilHelpers.Lambda((in Action action, in BitmapSourcesLinkerLoadStep _loadStep) => Run(action, _loadStep), (ref BitmapSourcesLinkerLoadStep __loadStep, ref BitmapSource __value) =>
+        {
+            IBitmapSources bitmapSources = GetBitmapSources(__loadStep);
 
-        public virtual void Load() => InvokeMethods(name => name.Length > nameof(Load).Length && name.StartsWith(nameof(Load)));
+            if (bitmapSources != null)
+
+                UpdateBitmapSource(ref __value, selector(bitmapSources));
+
+            __loadStep++;
+        }, ref loadStep, ref value);
+
+        protected IBitmapSources GetBitmapSources(BitmapSourcesLinkerLoadStep loadStep)
+        {
+            switch (loadStep)
+            {
+                case BitmapSourcesLinkerLoadStep.Default:
+
+                    return InnerObject.Intermediate;
+
+                case BitmapSourcesLinkerLoadStep.Intermediate:
+
+                    return InnerObject.Sources;
+
+                default:
+
+                    throw new InvalidOperationException("All bitmap sources have already been loaded.");
+            }
+        }
+
+        public void LoadSmall() => Run2(ref _bitmapSources.Small, bitmapSources => bitmapSources.Small, ref _loadSteps._loadSteps.Small);
+
+        public void LoadMedium() => Run2(ref _bitmapSources.Medium, bitmapSources => bitmapSources.Medium, ref _loadSteps._loadSteps.Medium);
+
+        public void LoadLarge() => Run2(ref _bitmapSources.Large, bitmapSources => bitmapSources.Large, ref _loadSteps._loadSteps.Large);
+
+        public void LoadExtraLarge() => Run2(ref _bitmapSources.ExtraLarge, bitmapSources => bitmapSources.ExtraLarge, ref _loadSteps._loadSteps.ExtraLarge);
+
+        public virtual void Load() => Run(() => InvokeMethods(name => name.Length > nameof(Load).Length && name.StartsWith(nameof(Load))), null);
 
         private void InvokeMethods(Predicate<string> predicate)
         {