+++ Source/IO/WinCopies.IO.Shell.Shared/BrowsableObjectInfo.cs
@@ -60,7 +60,7 @@ namespace WinCopies.IO.Shell
 
             public static Icon TryGetIcon(in Icon icon, in ushort size) => TryGetIcon(icon, new System.Drawing.Size(size, size));
 
-            public static Icon TryGetIcon(in Icon[] icons, in System.Drawing.Size size) => icons?.TryGetIcon(size, 32, true, true);
+            public static Icon TryGetIcon(in Icon[] icons, in System.Drawing.Size size) => icons?.TryGetIcon(size, 64, true, true);
 
             public static Icon TryGetIcon(in Icon icon, in System.Drawing.Size size) => TryGetIcon(icon?.Split(), size);